requirements.txt
blinker==1.8.2
click==8.1.7
coverage==7.6.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.6
